ai_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
chromadb
tavily-python
openai
python-multipart
requests
python-jose[cryptography]
passlib[bcrypt]
pypdf python-docx